Rename app page and decrease font weight in buy form
