+++ public/apps/index.html
@@ -57,7 +57,7 @@
             <li class="nav-item">
               <a class="nav-link" href="https://github.com/autonomoussoftware/documentation/blob/master/FAQ.md" target="_blank">FAQ</a>
             </li>
-            <li class="nav-item" x`id="auctions-menu-item" style="display: none;">
+            <li class="nav-item" id="auctions-menu-item" style="display: none;">
               <a class="nav-link" href="/auction">Auction</a>
             </li>
           </ul>
