+++ src/css/styles.less
@@ -1524,7 +1524,7 @@ a.overview__current-auction-error-link {
     top: -4px;
     padding: 0px;
     font-size: 24px;
-    font-weight: bold;
+    font-weight: 600;
     font-style: normal;
     font-stretch: normal;
     line-height: normal;
@@ -1615,7 +1615,7 @@ a.overview__current-auction-error-link {
     color: #fff;
     background-color: #7e61f8;
     font-size: 14px;
-    font-weight: bold;
+    font-weight: 600;
     font-style: normal;
     font-stretch: normal;
     line-height: 1.14;
@@ -1795,7 +1795,7 @@ a.overview__current-auction-error-link {
     color: @white;
     background-color: @purple;
     font-size: 14px;
-    font-weight: bold;
+    font-weight: 600;
     font-style: normal;
     font-stretch: normal;
     line-height: 1.14;
